refactor: move menu to its own file, rename options file
