+++ NodeToPython/ntp_menu.py
@@ -1,0 +1,13 @@
+import bpy
+
+class NodeToPythonMenu(bpy.types.Menu):
+    bl_idname = "NODE_MT_node_to_python"
+    bl_label = "Node To Python"
+
+    @classmethod
+    def poll(cls, context):
+        return True
+
+    def draw(self, context):
+        layout = self.layout.column_flow(columns=1)
+        layout.operator_context = 'INVOKE_DEFAULT'
